ekf2_test: create helper functions for auto-generated code diff
